+++ addons/point_of_sale/static/src/app/debug/debug_service.js
@@ -2,7 +2,7 @@
 
 import { reactive } from "@odoo/owl";
 import { registry } from "@web/core/registry";
-import { withComputedProperties } from "@point_of_sale/utils";
+import { withComputedProperties } from "@web/core/utils/reactive";
 import { DebugWidget } from "@point_of_sale/app/debug/debug_widget";
 
 export const debugService = {
